Try to make more sane file nameing and enums/structures placing.
