+++ src/config/options.rs
@@ -1,9 +1,48 @@
-use std::str::FromStr;
-
+use crate::ToDoError;
+use clap::ValueEnum;
 use serde::{Deserialize, Serialize};
+use std::fmt::Display;
+use std::str::FromStr;
 use tui::style::Modifier;
 
-use crate::ToDoError;
+#[derive(Serialize, Deserialize, ValueEnum, Clone, Debug, PartialEq, Eq, Default)]
+pub enum SetFinalDateType {
+    Override,
+    #[default]
+    OnlyMissing,
+    Never,
+}
+
+impl Display for SetFinalDateType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(&super::parsers::enum_debug_display_parser(&format!(
+            "{:?}",
+            self
+        )))?;
+        Ok(())
+    }
+}
+
+/// Represents the possible sorting options for tasks.
+#[derive(Clone, Copy, Serialize, Deserialize, Default, ValueEnum, Debug, PartialEq, Eq)]
+pub enum TaskSort {
+    #[default]
+    None,
+    Reverse,
+    Priority,
+    Alphanumeric,
+    AlphanumericReverse,
+}
+
+impl Display for TaskSort {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(&super::parsers::enum_debug_display_parser(&format!(
+            "{:?}",
+            self
+        )))?;
+        Ok(())
+    }
+}
 
 /// Serialization and deserialization support for the TUI text modifier type.
 ///
